% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{af_distribution}
\alias{af_distribution}
\title{Function that plots a histogram of the the variant allelic
support (tumor)}
\usage{
af_distribution(var_df, bin_size = 0.05)
}
\arguments{
\item{var_df}{data frame with somatic mutations}

\item{bin_size}{size of bins for allelic frequency}
}
\value{
p geom_histogram plot from ggplot2
}
\description{
Function that plots a histogram of the the variant allelic
support (tumor)
}
