% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_annotation_links}
\alias{append_annotation_links}
\title{Function that appends multiple HTML annotation links to variant identifiers
e.g. COSMIC, CLINVAR, REFSEQ etc}
\usage{
append_annotation_links(var_data_df, vartype = "snv_indel", skip = NULL)
}
\arguments{
\item{var_data_df}{data frame with variant entries}

\item{vartype}{'snv_indel' or 'cna', or 'exp'}

\item{skip}{elements to be ignored during annotation}
}
\description{
Function that appends multiple HTML annotation links to variant identifiers
e.g. COSMIC, CLINVAR, REFSEQ etc
}
