% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_dbmts_var_link}
\alias{append_dbmts_var_link}
\title{Function that adds miRNA target annotations (dbMTS) to
genetic variant identifiers}
\usage{
append_dbmts_var_link(var_df)
}
\arguments{
\item{var_df}{data frame with variants}
}
\description{
Function that adds miRNA target annotations (dbMTS) to
genetic variant identifiers
}
