% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_oncogenicity_docs}
\alias{append_oncogenicity_docs}
\title{Function that adds oncogenicity documentation from codes}
\usage{
append_oncogenicity_docs(var_df, ref_data = NULL)
}
\arguments{
\item{var_df}{Data frame of sample variants from VCF}

\item{ref_data}{DGG Engine reference data object}
}
\value{
vcf_data_df
}
\description{
Function that adds oncogenicity documentation from codes
}
