% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{append_targeted_drug_annotations}
\alias{append_targeted_drug_annotations}
\title{Function that adds link to targeted drugs (on and off-label)
for a list of variants and associated targeted}
\usage{
append_targeted_drug_annotations(
  var_df,
  primary_site = "Breast",
  ref_data = NULL
)
}
\arguments{
\item{var_df}{data frame with variants}

\item{primary_site}{primary site of tumor sample}

\item{ref_data}{DGG Engine reference data object}
}
\description{
Function that adds link to targeted drugs (on and off-label)
for a list of variants and associated targeted
}
