% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{assign_somatic_germline_evidence}
\alias{assign_somatic_germline_evidence}
\title{Function that appends several tags denoting
evidence for somatic/germline status of variants}
\usage{
assign_somatic_germline_evidence(sample_calls, settings = NULL)
}
\arguments{
\item{sample_calls}{data frame with variants}

\item{settings}{PCGR run/configuration settings}
}
\value{
sample_calls
}
\description{
Function that appends several tags denoting
evidence for somatic/germline status of variants
}
