% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomarker_evidence}
\alias{biomarker_evidence}
\title{Fixed data types/categories used for biomarker evidence, e.g. 'types','levels' etc.}
\format{
An object of class \code{list} of length 1.
}
\usage{
biomarker_evidence
}
\description{
Fixed data types/categories used for biomarker evidence, e.g. 'types','levels' etc.
}
\keyword{datasets}
