% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_common_colnames}
\alias{check_common_colnames}
\title{Function that checks whether a set of column names are present in two
different data frames}
\usage{
check_common_colnames(df1 = NULL, df2 = NULL, cnames = NULL)
}
\arguments{
\item{df1}{data frame 1}

\item{df2}{data frame 2}

\item{cnames}{character vector with column names to be check for presence}
}
\value{
existing_common_columns T/F
}
\description{
Function that checks whether a set of column names are present in two
different data frames
}
