% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_palette}
\alias{color_palette}
\title{Color encodings for report elements of DGG Engine/CPSR}
\format{
A list object with different report elements that are color-coded in
DGG Engine/CPSR reports. Each list element have two vectors: 'levels' and 'values'.
Currently, the following list elements are included:
\itemize{
\item \emph{pathogenicity} - Colors for five-level pathogenicity levels (CPSR)
\item \emph{clinical_evidence} - Colors for strength of evidence of cancer-variant associations (A-E)
\item \emph{tier} - Colors for tier levels for variant prioritization (DGG Engine)
\item \emph{report_color} - Colors for DGG Engine assay mode (tumor-control vs. tumor-only)
\item \emph{warning} - Color for warning (low confidence in DGG Engine analysis output)
\item \emph{success} - Color for success (no evident uncertainty in DGG Engine analysis output)
}
}
\usage{
color_palette
}
\description{
Color encodings for report elements of DGG Engine/CPSR
}
\keyword{datasets}
