% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{expand_biomarker_items}
\alias{expand_biomarker_items}
\title{Function that expands biomarker evidence items with variant annotations}
\usage{
expand_biomarker_items(
  callset = NULL,
  variant_origin = "somatic",
  target_genes = NULL
)
}
\arguments{
\item{callset}{list object with 'variant' and 'biomarker_evidence' data
frames}

\item{variant_origin}{'somatic' or 'germline'}

\item{target_genes}{data frame with target genes of interest}
}
\description{
Function that expands biomarker evidence items with variant annotations
}
