% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkers.R
\name{get_clin_assocs_cna}
\alias{get_clin_assocs_cna}
\title{Function that retrieves clinical evidence items (CIVIC, CBMDB) for
CNA aberrations}
\usage{
get_clin_assocs_cna(onco_ts_sets, annotation_tags = NULL, eitems = NULL)
}
\arguments{
\item{onco_ts_sets}{data frame with annotations with respect to
lost tumor suppressor genes and gained oncogenes}

\item{annotation_tags}{list with annotation tags for display in report}

\item{eitems}{Data frame with clinical evidence items}
}
\value{
list
}
\description{
Function that retrieves clinical evidence items (CIVIC, CBMDB) for
CNA aberrations
}
