% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_data.R
\name{get_excel_sheets}
\alias{get_excel_sheets}
\title{Function that produces the contents of sheets for an Excel report
of PCGR output}
\usage{
get_excel_sheets(report = NULL)
}
\arguments{
\item{report}{PCGR report object}
}
\description{
Function that produces the contents of sheets for an Excel report
of PCGR output
}
