% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune_celltypes}
\alias{immune_celltypes}
\title{Data frame with immune cell types}
\format{
A data.frame with 11 rows and 2 columns that indicates immune
cell types used in immune contexture analysis by quanTIseq
}
\usage{
immune_celltypes
}
\description{
Data frame with immune cell types
}
\keyword{datasets}
