% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_msi_content}
\alias{init_msi_content}
\title{Function that initiates report element with MSI classification}
\usage{
init_msi_content()
}
\description{
Function that initiates report element with MSI classification
}
