% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_rainfall_content}
\alias{init_rainfall_content}
\title{Function that initiates report element with rainfall information}
\usage{
init_rainfall_content()
}
\value{
rep Report structure initialized for rainfall data
}
\description{
Function that initiates report element with rainfall information
}
