% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{init_tumor_only_content}
\alias{init_tumor_only_content}
\title{Function that initiates report element with tumor-only information}
\usage{
init_tumor_only_content()
}
\value{
rep Report structure initialized for tumor-only stats
}
\description{
Function that initiates report element with tumor-only information
}
