% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_expression_csq}
\alias{load_expression_csq}
\title{Load expression consequence settings}
\usage{
load_expression_csq(settings = NULL)
}
\arguments{
\item{settings}{PCGR run/configuration settings}
}
\description{
Load expression consequence settings
}
