% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_expression_similarity}
\alias{load_expression_similarity}
\title{Load expression similarity results}
\usage{
load_expression_similarity(settings = NULL)
}
\arguments{
\item{settings}{PCGR run/configuration settings}
}
\description{
Load expression similarity results
}
