% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_somatic_cna}
\alias{load_somatic_cna}
\title{Function that reads and validates fully annotated CNA data (segments and genes)
from PCGR pre-reporting pipeline}
\usage{
load_somatic_cna(
  fname_cna_segment = NULL,
  fname_cna_gene = NULL,
  ref_data = NULL,
  settings = NULL
)
}
\arguments{
\item{fname_cna_segment}{Path to file with pre-processed CNA segments}

\item{fname_cna_gene}{Path to file with pre-processed CNA gene-level data}

\item{ref_data}{PCGR reference data object}

\item{settings}{PCGR run/configuration settings}
}
\description{
Function that reads and validates fully annotated CNA data (segments and genes)
from PCGR pre-reporting pipeline
}
