% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log4r_info}
\alias{log4r_info}
\title{Write messages to logs at a given priority level}
\usage{
log4r_info(msg)
}
\arguments{
\item{msg}{Message to log.}
}
\description{
See \code{\link[log4r:levellog]{log4r::levellog()}}
}
