% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_boxes.R
\name{plot_value_boxes}
\alias{plot_value_boxes}
\title{Function that plots four value boxes with the most
important findings in the cancer genome}
\usage{
plot_value_boxes(pcg_report)
}
\arguments{
\item{pcg_report}{pcg report with list elements}
}
\value{
p
}
\description{
Function that plots four value boxes with the most
important findings in the cancer genome
}
