% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_cols_from_df}
\alias{remove_cols_from_df}
\title{Function that removes column(s) from data frame}
\usage{
remove_cols_from_df(df, cnames = NULL)
}
\arguments{
\item{df}{data.frame with data}

\item{cnames}{character vector with column names}
}
\value{
df data.frame with columns removed (if present originally)
}
\description{
Function that removes column(s) from data frame
}
