% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_cohorts}
\alias{tcga_cohorts}
\title{Data frame with all TCGA cohorts}
\format{
A data.frame with 33 rows and 2 columns that indicates TCGA cohorts
}
\usage{
tcga_cohorts
}
\description{
Data frame with all TCGA cohorts
}
\keyword{datasets}
