% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline.R
\name{tcga_somatic_status}
\alias{tcga_somatic_status}
\title{Function that assigns a logical (STATUS_TCGA_SOMATIC) reflecting whether
a variant co-incides with an entry in TCGA (somatic)}
\usage{
tcga_somatic_status(sample_calls)
}
\arguments{
\item{sample_calls}{data frame with sample variant calls}
}
\description{
Function that assigns a logical (STATUS_TCGA_SOMATIC) reflecting whether
a variant co-incides with an entry in TCGA (somatic)
}
