% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_processed_vcf}
\alias{write_processed_vcf}
\title{Function that writes a VCF intended for mutational signature analysis}
\usage{
write_processed_vcf(
  calls,
  sample_name = NULL,
  output_directory = NULL,
  vcf_fname = NULL,
  snv_only = TRUE
)
}
\arguments{
\item{calls}{data frame with calls}

\item{sample_name}{sample name}

\item{output_directory}{Output directory for output file}

\item{vcf_fname}{filename for VCF}

\item{snv_only}{logical, if TRUE only SNVs are written to VCF}
}
\description{
Function that writes a VCF intended for mutational signature analysis
}
