% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_report_excel}
\alias{write_report_excel}
\title{Function that writes key datasets of PCGR object to an Excel workbook}
\usage{
write_report_excel(report = NULL)
}
\arguments{
\item{report}{List object with all PCGR report data, settings etc.}
}
\description{
Function that writes key datasets of PCGR object to an Excel workbook
}
