% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_report_quarto_html}
\alias{write_report_quarto_html}
\title{Function that writes contents of PCGR object to an HTML report (quarto-based)}
\usage{
write_report_quarto_html(report = NULL)
}
\arguments{
\item{report}{List object with all report data, settings etc.}
}
\description{
Function that writes contents of PCGR object to an HTML report (quarto-based)
}
